+++ src/main/kotlin/net/pwall/json/JSONDeserializer.kt
@@ -32,15 +32,19 @@ import kotlin.reflect.KClass
 import kotlin.reflect.KFunction
 import kotlin.reflect.KMutableProperty
 import kotlin.reflect.KParameter
+import kotlin.reflect.KProperty
 import kotlin.reflect.KType
 import kotlin.reflect.KTypeProjection
 import kotlin.reflect.full.*
+import kotlin.reflect.jvm.isAccessible
 
+import java.time.Duration
 import java.time.Instant
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.OffsetDateTime
 import java.time.OffsetTime
+import java.time.Period
 import java.time.Year
 import java.time.YearMonth
 import java.time.ZonedDateTime
@@ -215,10 +219,17 @@ object JSONDeserializer {
 
                 YearMonth::class -> return YearMonth.parse(str) as T
 
+                Duration::class -> return Duration.parse(str) as T
+
+                Period::class -> return Period.parse(str) as T
+
                 UUID::class -> return UUID.fromString(str) as T
 
             }
 
+        }
+        catch (e: JSONException) {
+            throw e
         }
         catch (e: Exception) {
             throw JSONException("Can't deserialize \"$str\" as $resultClass", e)
@@ -331,6 +342,9 @@ object JSONDeserializer {
             } as T
 
         }
+        catch (e: JSONException) {
+            throw e
+        }
         catch (e: Exception) {
             throw JSONException("Can't deserialize array as $resultClass", e)
         }
@@ -355,6 +369,8 @@ object JSONDeserializer {
                 return (if (isMutable(resultClass)) result.toMutableMap() else result.toMap()) as T
             }
 
+            resultClass.objectInstance?.let { return setRemainingFields(resultClass, it, json, config) }
+
             findBestConstructor(resultClass.constructors, json)?.let { constructor ->
                 val argMap = HashMap<KParameter, Any?>()
                 val jsonCopy = HashMap<String, JSONValue?>(json)
@@ -365,13 +381,7 @@ object JSONDeserializer {
                         jsonCopy.remove(paramName)
                     }
                 }
-                val instance = constructor.callBy(argMap)
-                jsonCopy.forEach { entry -> // JSONObject fields not used in constructor
-                    val member = findFieldToSet(resultClass.members, entry.key) ?:
-                            throw JSONException("Can't find property ${entry.key} in ${resultClass.simpleName}")
-                    member.setter.call(instance, deserialize(member.returnType, jsonCopy[entry.key], config))
-                }
-                return instance
+                return setRemainingFields(resultClass, constructor.callBy(argMap), jsonCopy, config)
             }
         }
         catch (e: JSONException) {
@@ -385,9 +395,36 @@ object JSONDeserializer {
 
     }
 
-    private fun findFieldToSet(members: Collection<KCallable<*>>, name: String): KMutableProperty<*>? {
+    private fun <T: Any>  setRemainingFields(resultClass: KClass<T>, instance: T, json: Map<String, JSONValue?>,
+            config: JSONConfig?): T {
+        json.forEach { entry -> // JSONObject fields not used in constructor
+            val member = findField(resultClass.members, entry.key) ?:
+                    throw JSONException("Can't find property ${entry.key} in ${resultClass.simpleName}")
+            val value = deserialize(member.returnType, json[entry.key], config)
+            if (member is KMutableProperty<*>) {
+                val wasAccessible = member.isAccessible
+                member.isAccessible = true
+                try {
+                    member.setter.call(instance, value)
+                }
+                catch (e: Exception) {
+                    throw JSONException("Error setting property ${entry.key} in ${resultClass.simpleName}", e)
+                }
+                finally {
+                    member.isAccessible = wasAccessible
+                }
+            }
+            else {
+                if (member.getter.call(instance) != value)
+                    throw JSONException("Can't set property ${entry.key} in ${resultClass.simpleName}")
+            }
+        }
+        return instance
+    }
+
+    private fun findField(members: Collection<KCallable<*>>, name: String): KProperty<*>? {
         for (member in members) {
-            if (member is KMutableProperty<*> && member.aName() == name)
+            if (member is KProperty<*> && member.aName() == name)
                     return member
         }
         return null
@@ -426,7 +463,7 @@ object JSONDeserializer {
         return findAnnotation<JSONName>()?.name ?: name
     }
 
-    private fun KMutableProperty<*>.aName(): String? { // annotated name, or regular name if not overridden
+    private fun KProperty<*>.aName(): String? { // annotated name, or regular name if not overridden
         return findAnnotation<JSONName>()?.name ?: name
     }