+++ src/test/kotlin/net/pwall/json/JSONSerializerTest.kt
@@ -31,11 +31,13 @@ import kotlin.test.assertNull
 import kotlin.test.assertSame
 import kotlin.test.assertTrue
 
+import java.time.Duration
 import java.time.Instant
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.OffsetDateTime
 import java.time.OffsetTime
+import java.time.Period
 import java.time.Year
 import java.time.YearMonth
 import java.time.ZoneId
@@ -323,6 +325,18 @@ class JSONSerializerTest {
         assertEquals(expected, JSONSerializer.serialize(year))
     }
 
+    @Test fun `Duration should return JSONString`() {
+        val duration = Duration.ofHours(2)
+        val expected = JSONString("PT2H")
+        assertEquals(expected, JSONSerializer.serialize(duration))
+    }
+
+    @Test fun `Period should return JSONString`() {
+        val period = Period.ofMonths(3)
+        val expected = JSONString("P3M")
+        assertEquals(expected, JSONSerializer.serialize(period))
+    }
+
     @Test fun `UUID should return JSONString`() {
         val uuidString = "12ce3730-2d97-11e7-aeed-67b0e6bf0ed7"
         val uuid = UUID.fromString(uuidString)
@@ -404,6 +418,31 @@ class JSONSerializerTest {
         val triple = Triple("xyz",88,"def")
         val expected = JSONArray().addJSON(JSONString("xyz")).addJSON(JSONInteger(88)).addJSON(JSONString("def"))
         assertEquals(expected, JSONSerializer.serialize(triple))
+    }
+
+    @Test fun `object should return JSONObject()`() {
+        val obj = DummyObject
+        val expected = JSONObject().putValue("field1", "abc")
+        assertEquals(expected, JSONSerializer.serialize(obj))
+    }
+
+    @Test fun `nested object should return JSONObject()`() {
+        val obj = NestedDummy()
+        val nested = JSONObject().putValue("field1", "abc")
+        val expected = JSONObject().putJSON("obj", nested)
+        assertEquals(expected, JSONSerializer.serialize(obj))
+    }
+
+    @Test fun `class with constant val should serialize correctly`() {
+        val constClass = DummyWithVal()
+        val expected = JSONObject().putValue("field8", "blert")
+        assertEquals(expected, JSONSerializer.serialize(constClass))
+    }
+
+    @Test fun `java class should serialize correctly`() {
+        val javaClass1 = JavaClass1(1234, "Hello!")
+        val expected = JSONObject().putValue("field1", 1234).putValue("field2", "Hello!")
+        assertEquals(expected, JSONSerializer.serialize(javaClass1))
     }
 
     private fun intEquals(a: Int, b: Int): Boolean {